lmsensors: Mark as only buildable when MACHINE is set
Requiring kernel-dev requires a kernel provider (recipe) to be available,
and this is normally not the case for builds where MACHINE is not set.
If MACHINE is set, and no kernel provider is available, this is not
enough and lmsensors recipe will be left as broken, resulting in
"FATAL: No provider for machine:kernel-dev (needed by
machine:lmsensors_3.3.3)" when trying to build it.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>